# openapi.yaml
openapi: 3.0.3
info:
  title: eTutor Task-App API Contract
  description: >-
    This document describes the RESTful API that eTutor Task apps must comply
    with in order for task creation to work with eTutor Task-Administration.
    Task apps may have additional endpoints that can be called by the task administration UI.
    The task administration server will forward the requests from the UI to the task app.

    <p>Authentication is performed using API keys. Different API keys may have
    different roles:</p>

    <ul>
        <li><em>CRUD</em>: Create, read, update, and delete tasks and task groups.</li>
        <li><em>SUBMIT</em>: Submit submissions</li>
        <li><em>READ_SUBMISSION</em>: Read persisted submissions</li>
    </ul>
  version: 1.1.1
...
paths:
  /api/taskGroup/{id}:
    get:
      tags:
        - TaskGroup
      operationId: getTaskGroup
      summary: Get task group
      description: >-
        Returns the task group type specific data of the requested task group.
        This is used by the task administration to retrieve the form values for
        editing the task group. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task group identifier.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroup'
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
        '404':
          description: Task group not found
      security:
        - api_key: [ ]
    post:
      tags:
        - TaskGroup
      operationId: createTaskGroup
      summary: Create task group
      description: >-
        Creates a new task group. The new task group is stored under the
        specified identifier. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task group identifier.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Task group data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyTaskGroup'
        required: true
      responses:
        '201':
          description: Task group created
          headers:
            Location:
              description: The location of the created task group.
              schema:
                example: /api/taskGroup/2
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupModificationResponse'
        '400':
          description: Invalid task group data
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
      security:
        - api_key: [ ]
    put:
      tags:
        - TaskGroup
      operationId: updateTaskGroup
      summary: Modify task group
      description: Updates an existing task group. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task group identifier.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Task group data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyTaskGroup'
        required: true
      responses:
        '200':
          description: Task group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupModificationResponse'
        '204':
          description: Task group updated and response body is empty
        '400':
          description: Invalid task group data
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
        '404':
          description: Task group not found
      security:
        - api_key: [ ]
    delete:
      tags:
        - TaskGroup
      operationId: deleteTaskGroup
      summary: Delete task group
      description: >-
        Deletes an existing task group. Does not return an error if the
        specified task group does not exist. Requires the <code>CRUD</code>
        role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task group identifier.
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task group deleted
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
      security:
        - api_key: [ ]
  /api/task/{id}:
    get:
      tags:
        - Task
      operationId: getTask
      summary: Get task
      description: >-
        Returns the task type specific data of the requested task. This is used
        by the task administration to retrieve the form values for editing the
        task. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task identifier.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
        '404':
          description: Task not found
      security:
        - api_key: [ ]
    post:
      tags:
        - Task
      operationId: createTask
      summary: Create task
      description: >-
        Creates a new task. The new task is stored under the specified
        identifier. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task identifier.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Task data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyTask'
        required: true
      responses:
        '201':
          description: Task created
          headers:
            Location:
              description: The location of the created task.
              schema:
                example: /api/task/2
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskModificationResponse'
        '400':
          description: Invalid task data
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
      security:
        - api_key: [ ]
    put:
      tags:
        - Task
      operationId: updateTask
      summary: Modify task
      description: Updates an existing task. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task identifier.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Task data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyTask'
        required: true
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskModificationResponse'
        '204':
          description: Task updated and response body is empty
        '400':
          description: Invalid task data
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
        '404':
          description: Task not found
      security:
        - api_key: [ ]
    delete:
      tags:
        - Task
      operationId: deleteTask
      summary: Delete task
      description: >-
        Deletes an existing task. Does not return an error if the specified task
        does not exist. Requires the <code>CRUD</code> role.
      parameters:
        - name: id
          in: path
          required: true
          description: The task identifier.
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task deleted
        '401':
          description: Unauthorized
        '403':
          description: Operation not allowed
      security:
        - api_key: [ ]
...
components:
  schemas:
    TaskGroup:
      type: object
      description: >-
        The task group type specific data of a task group. The content of this
        object depends on the task group type.
      additionalProperties: true
      example:
        ddl: CREATE TABLE test (id INTEGER);
    ModifyTaskGroup:
      description: Data for creating and updating task groups.
      type: object
      required:
        - taskGroupType
        - additionalData
        - status
      properties:
        taskGroupType:
          type: string
          minLength: 2
          maxLength: 100
          description: >-
            The task group type. If a task app supports more than one task group
            type, this can be used to distinguish between types.
          example: sql
        status:
          type: string
          description: The task group status.
          enum:
            - DRAFT
            - READY_FOR_APPROVAL
            - APPROVED
        additionalData:
          type: object
          additionalProperties: true
          description: >-
            The task group type specific data. The content of this object
            depends on the task group type.
          example:
            ddl: CREATE TABLE test (id INTEGER);
    TaskGroupModificationResponse:
      description: Response data for creating and updating task groups. The task administration will use this value to set the task group description, if the user did not enter a description.
      type: object
      properties:
        descriptionDe:
          type: string
          description: The german task group description. May be <code>null</code> if generation of task group description is not supported.
          example: "Gesucht ist eine Zahl zwischen 1 und 10."
        descriptionEn:
          type: string
          description: The english task group description. May be <code>null</code> if generation of task group description is not supported.
          example: "We are looking for a number between 1 and 10."
    Task:
      type: object
      description: >-
        The task type specific data of a task. The content of this object
        depends on the task type.
      additionalProperties: true
      example:
        solution: SELECT id FROM student;
    ModifyTask:
      description: Data for creating and updating tasks
      type: object
      required:
        - additionalData
        - maxPoints
        - status
        - taskType
      properties:
        taskGroupId:
          type: integer
          description: The task group identifier, if applicable.
          format: int64
          example: 445
        maxPoints:
          type: number
          description: The maximum achievable points.
          example: 10.5
        taskType:
          maxLength: 100
          minLength: 2
          type: string
          description: >-
            The task type. If a task app supports more than one task type, this
            can be used to distinguish between types.
          example: sql
        status:
          type: string
          description: The task status.
          enum:
            - DRAFT
            - READY_FOR_APPROVAL
            - APPROVED
        additionalData:
          type: object
          additionalProperties: true
          description: >-
            The task type specific data. The content of this object depends on
            the task type.
          example:
            solution: SELECT id FROM student;
    TaskModificationResponse:
      description: Response data for creating and updating tasks.
      type: object
      properties:
        descriptionDe:
          type: string
          description: The german task description. May be <code>null</code> if generation of task description is not supported. The task administration will use this value to set the task description, if the user did not enter a description.
          example: "Geben Sie eine Zahl ein."
        descriptionEn:
          type: string
          description: The english task description. May be <code>null</code> if generation of task description is not supported. The task administration will use this value to set the task description, if the user did not enter a description.
          example: "Enter a number."
        difficulty:
          type: number
          format: int32
          description: The difficulty of the task. May be <code>null</code>. If a value is set, the task administration will override the user-specified difficulty with this value. Might be used by task apps to automatically set difficulty based on some metric.
          example: 2
          minimum: 0
          maximum: 3
        maxPoints:
          type: number
          description: The maximum points. May be <code>null</code>. If a value is set, the task administration will override the user-specified maximum points with this value. Might be used by task apps to automatically set maximum points based on some grading scheme.
          example: 25.75
          minimum: 0.01
...
  securitySchemes:
    api_key:
      description: Authentication with API Key
      type: apiKey
      name: X-API-KEY
      in: header